improve ApplicableTradeTax
